chore: refactoring metadata storage and deduping tests
