+++ src/index.ts
@@ -2,7 +2,7 @@ export * from './Decorators';
 export * from './BaseFirestoreRepository';
 export * from './types';
 export * from './helpers';
-export { initialize, Initialize } from './MetadataStorage';
+export { initialize, Initialize } from './MetadataUtils';
 
 // Temporary while https://github.com/wovalle/fireorm/issues/58 is being fixed
 export { Type } from 'class-transformer';
